Bug/SU-615: Regenerate accrual transaction on maximum legal rate change
