+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/jobs/dailyaccrual/DailyInterestAccrualPosterTask.java
@@ -47,10 +47,6 @@ public class DailyInterestAccrualPosterTask implements Callable<Void> {
         this.dailyInterestAccrualPoster.setLoanIds(loanIds);
     }
 
-    public void setMinimumDaysInArrearsToSuspendLoanAccount(Long minimumDaysInArrearsToSuspendLoanAccount) {
-        this.dailyInterestAccrualPoster.setMinimumDaysInArrearsToSuspendLoanAccount(minimumDaysInArrearsToSuspendLoanAccount);
-    }
-
     public void setAccrualDate(LocalDate accrualDate) {
         this.dailyInterestAccrualPoster.setAccrualDate(accrualDate);
     }
