+++ fineract-provider/src/main/java/org/apache/fineract/portfolio/loanaccount/service/LoanAccrualPlatformService.java
@@ -32,4 +32,6 @@ public interface LoanAccrualPlatformService {
 
     void addPeriodicAccruals(LocalDate tilldate, Collection<LoanScheduleAccrualData> loanScheduleAccrualDatas) throws MultiException;
 
+    void persistDailyInterestAccrual(Long loanId, LocalDate accrualDate);
+
 }
